requirements.txt
openai
langchain_openai
langgraph 
langchain-communit
langchain
langchain-classic


